Cogs branch rewrite
